Move AstroBin integration to its own dedicated settings tab
